Skip over invalid vine URLs.
In http://inkdroid.org/vine/20161105105755.txt we have a line that looks
like

	794852785258975233 https://vine.co/v/eHpJlZdrdmY---KENDÄ°

Given this text, URI will raise

	URI::InvalidURIError: URI must be ascii only "https://vine.co/v/eHpJlZdrdmY---KEND\xC4\xB0"
	  /usr/local/lib/ruby/2.3.0/uri/rfc3986_parser.rb:21:in `split'
	  /usr/local/lib/ruby/2.3.0/uri/rfc3986_parser.rb:73:in `parse'
	  /usr/local/lib/ruby/2.3.0/uri/common.rb:227:in `parse'
	  /usr/local/lib/ruby/2.3.0/uri/common.rb:714:in `URI'
	  /code/lib/inkdroid/jobs.rb:53:in `block in perform'

Our goal is to ingest as much as possible with as little human oversight
as possible.  Best fit solution: expect that we'll see that Vine URL
again via some other process and move on to the next one.